Move ignore rules parsing into lib.rs
