+++ src/lib.rs
@@ -1,16 +1,18 @@
-use error::SpiderError;
+use error::{SpiderError, SpiderErrorType};
 
+use log::info;
 use petgraph::graph::{DiGraph, NodeIndex};
 use reqwest::StatusCode;
 use scraper::{selector::CssLocalName, Selector};
-use std::collections::HashMap;
+use std::io::{BufReader, BufRead};
+use std::str::FromStr;
+use std::{collections::HashMap, fs::File};
 use std::sync::Mutex;
 use url::{Host, Url};
 
 pub mod algo;
 pub mod dot;
 pub mod error;
-pub mod ignore_rules;
 pub mod url_helpers;
 
 #[cfg(test)]
@@ -81,6 +83,8 @@ pub struct SpiderOptions {
     pub skip_class: CssLocalName,
     /// Vector of hosts (domain names and IP addresses) that Spider Crab will traverse
     pub hosts: Vec<Host<String>>,
+    /// List of patterns that the user has specified to ignore
+    pub ignore_patterns: HashMap<SpiderErrorType, Vec::<String>>
 }
 
 impl SpiderOptions {
@@ -100,6 +104,49 @@ impl SpiderOptions {
         self.hosts
             .push(Url::parse(url).unwrap().host().unwrap().to_owned())
     }
+
+    pub fn is_rule_enabled(&self, rule: SpiderErrorType, url: &Url) -> bool {
+        let patterns = self.ignore_patterns.get(&rule);
+        if patterns.is_none() {
+            return true;
+        }
+        let patterns = patterns.unwrap();
+        for p in patterns {
+            if p == url.as_str() {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    pub fn read_ignore_list_from_file(&mut self, filepath: &str) {
+        let ignore_file = File::open(filepath).unwrap();
+        let reader = BufReader::new(ignore_file);
+        let mut line_num = 0;
+        let mut count = 0;
+        for line in reader.lines() {
+            line_num += 1;
+            if let Ok(line) = line {
+                let line = line.trim();
+                if line.starts_with("#") || line.is_empty() {
+                    continue;
+                }
+                let mut parts = line.split_whitespace();
+                let rule = parts.next().expect(format!("Could not read ignore rule from line {} in the ignore file!", line_num).as_str());
+                let url = parts.next().expect(format!("Could not read URL from line {} in the ignore file!", line_num).as_str());
+                let error_type = SpiderErrorType::from_str(rule).expect(format!("Invalid ignore rule on line {} of ignore file!", line_num).as_str());
+
+                if ! self.ignore_patterns.contains_key(&error_type) {
+                    self.ignore_patterns.insert(error_type.clone(), Vec::new());
+                }
+
+                // The get_mut().unwrap() _should_ never panic because we check to make sure that the map contains the key right before this
+                self.ignore_patterns.get_mut(&error_type).unwrap().push(url.to_string());
+                count += 1;
+            }
+        }
+        info!("Parsed {} ignore rules from {}", count, filepath);
+    }
 }
 
 impl Default for SpiderOptions {
@@ -112,6 +159,7 @@ impl Default for SpiderOptions {
             title_selector: Box::new(Selector::parse("title").expect("Invalid <title> selector!")),
             skip_class: CssLocalName::from("scrab-skip"),
             hosts: vec![],
+            ignore_patterns: HashMap::new()
         }
     }
 }
